Refactor request optimizer to be more flexible (#726)
Refactor BrokerRequestOptimizer to be more flexible by allowing future
optimizations to be added in. Separated optimizations to have
optimizations that run on the filter tree and extracted the current
flattening optimization to be in a separate class.
