+++ pinot-transport/src/main/java/com/linkedin/pinot/requestHandler/BrokerRequestOptimizer.java
@@ -15,87 +15,49 @@
  */
 package com.linkedin.pinot.requestHandler;
 
-import java.util.ArrayList;
-import java.util.Iterator;
+import java.util.Arrays;
 import java.util.List;
 import com.linkedin.pinot.common.request.BrokerRequest;
-import com.linkedin.pinot.common.request.FilterOperator;
 import com.linkedin.pinot.common.request.FilterQuery;
 import com.linkedin.pinot.common.utils.request.FilterQueryTree;
 import com.linkedin.pinot.common.utils.request.RequestUtils;
 
 
 public class BrokerRequestOptimizer {
-  public static int MAX_OPTIMIZING_DEPTH = 5;
+  private static final List<? extends FilterQueryTreeOptimizer> FILTER_QUERY_TREE_OPTIMIZERS = Arrays.asList(
+      new FlattenNestedPredicatesFilterQueryTreeOptimizer()
+  );
+
   /**
+   * Optimizes the given broker request.
    *
-   * @param inputReq BrokerRequest that is to be optimized
-   * @return optimized request
+   * @param brokerRequest BrokerRequest that is to be optimized
+   * @return An optimized request
    */
-  public BrokerRequest optimize(BrokerRequest inputReq) {
-    // Flatten the query tree if possible
-    return flatten(inputReq);
-  }
+  public BrokerRequest optimize(BrokerRequest brokerRequest) {
+    optimizeFilterQueryTree(brokerRequest);
 
-  private BrokerRequest flatten(BrokerRequest request) {
-    FilterQueryTree root = null;
-    FilterQuery q = request.getFilterQuery();
-
-    if (q == null || request.getFilterSubQueryMap() == null) {
-      return request;
-    }
-
-    root = RequestUtils.buildFilterQuery(q.getId(), request.getFilterSubQueryMap().getFilterQueryMap());
-
-    flatten(root, null, MAX_OPTIMIZING_DEPTH);
-
-    RequestUtils.generateFilterFromTree(root, request);
-
-    return request;
+    return brokerRequest;
   }
 
   /**
-   * Flatten the operators if parent and child have the same AND or OR operator.
-   * (e.g. AND( a, AND (b, c)) is the same as AND(a, b, c). This helps when we re-order
-   * operators for performance.
-   *
-   * It does so by looking at the operator of the 'parent' and 'node'. If they are same, and
-   * collapsible, then all the children of 'node' are moved one level up to be siblings of
-   * 'node', rendering 'node' childless. 'node' is then removed from 'parent's children list.
-   *
-   * @param node The node whose children are to be moved up one level if criteria is satisfied.
-   * @param parent Node's parent who will inherit node's children if criteria is satisfied.
-   * @param maxDepth is the maximum depth to which we recurse
+   * Optimizes the filter query tree of a broker request in place.
+   * @param brokerRequest The broker request to optimize
    */
-  private void flatten(FilterQueryTree node, FilterQueryTree parent, int maxDepth) {
-    if (node == null || node.getChildren() == null || maxDepth == 0) {
+  private void optimizeFilterQueryTree(BrokerRequest brokerRequest) {
+    FilterQueryTree filterQueryTree = null;
+    FilterQuery q = brokerRequest.getFilterQuery();
+
+    if (q == null || brokerRequest.getFilterSubQueryMap() == null) {
       return;
     }
-    // Flatten all the children first.
-    List<FilterQueryTree> toFlatten = new ArrayList<>(node.getChildren().size());
-    for (FilterQueryTree child: node.getChildren()) {
-      if (child.getChildren() != null && !child.getChildren().isEmpty()) {
-        toFlatten.add(child);
-      }
+
+    filterQueryTree = RequestUtils.buildFilterQuery(q.getId(), brokerRequest.getFilterSubQueryMap().getFilterQueryMap());
+
+    for (FilterQueryTreeOptimizer filterQueryTreeOptimizer : FILTER_QUERY_TREE_OPTIMIZERS) {
+      filterQueryTree = filterQueryTreeOptimizer.optimize(filterQueryTree);
     }
-    for (FilterQueryTree child : toFlatten) {
-      flatten(child, node, maxDepth-1);
-    }
-    if (parent == null) {
-      return;
-    }
-    if (node.getOperator() == parent.getOperator() &&
-        (node.getOperator() == FilterOperator.OR || node.getOperator() == FilterOperator.AND)) {
-      // Move all of 'node's children one level up. If 'node' has no children left, remove it from parent's list.
-      List<FilterQueryTree> children = node.getChildren();
-      Iterator<FilterQueryTree> it = children.iterator();
-      while (it.hasNext()) {
-        parent.getChildren().add(it.next());
-        it.remove();
-      }
-      // 'node' is now childless
-      // Remove this node from its parent's list.
-      parent.getChildren().remove(node);
-    }
+
+    RequestUtils.generateFilterFromTree(filterQueryTree, brokerRequest);
   }
 }
